Move utility functions to a better location
